Исправила ошибки ДЗ Дебаг
