+++ src/main/java/me/matvienkoeg/recipeapp/controllers/FilesController.java
@@ -4,6 +4,7 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import me.matvienkoeg.recipeapp.services.FilesIngredientService;
 import me.matvienkoeg.recipeapp.services.FilesRecipeService;
+import me.matvienkoeg.recipeapp.services.RecipeService;
 import org.apache.commons.io.IOUtils;
 import org.springframework.core.io.InputStreamResource;
 import org.springframework.http.HttpHeaders;
@@ -17,16 +18,19 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.*;
+
 @RestController
 @Tag(name = "API по работе с файлами рецептов и ингредиентов и сохранение их")
 
 public class FilesController {
     private final FilesRecipeService filesRecipeService;
     private final FilesIngredientService filesIngredientService;
+    private final RecipeService recipeService;
 
-    public FilesController(FilesRecipeService filesRecipeService, FilesIngredientService filesIngredientService) {
+    public FilesController(FilesRecipeService filesRecipeService, FilesIngredientService filesIngredientService, RecipeService recipeService) {
         this.filesRecipeService = filesRecipeService;
         this.filesIngredientService = filesIngredientService;
+        this.recipeService = recipeService;
     }
 
     @GetMapping("/exportRecipe")
@@ -42,7 +46,7 @@ public class FilesController {
                     .contentLength(recipeFile.length())
                     .header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=" + recipeFile.getName())
                     .body(resource);
-        } catch (IOException e){
+        } catch (IOException e) {
             e.printStackTrace();
             return ResponseEntity.noContent().build();
         }
@@ -54,13 +58,13 @@ public class FilesController {
     )
     public ResponseEntity<InputStreamResource> downloadDataFileTxt() {
         try {
-            File recipeFile = filesRecipeService.prepareRecipesTxt();
+            File recipeFile = recipeService.prepareRecipesTxt();
             InputStreamResource resource = new InputStreamResource(new FileInputStream(recipeFile));
-                return ResponseEntity.ok()
-                        .contentType(MediaType.APPLICATION_OCTET_STREAM)
-                        .contentLength(recipeFile.length())
-                        .header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename="+ recipeFile.getName())
-                        .body(resource);
+            return ResponseEntity.ok()
+                    .contentType(MediaType.APPLICATION_OCTET_STREAM)
+                    .contentLength(recipeFile.length())
+                    .header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=" + recipeFile.getName())
+                    .body(resource);
 
         } catch (IOException e) {
             e.printStackTrace();
@@ -68,29 +72,29 @@ public class FilesController {
         }
     }
 
-        @PostMapping(value = "/importRecipe", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-        public ResponseEntity<String> uploadDataFileRecipe (@RequestParam MultipartFile file){
-            filesRecipeService.cleanDataFile();
-            File dataFile = filesRecipeService.getDataFile();
-            try (FileOutputStream fos = new FileOutputStream(dataFile)) {
-                IOUtils.copy(file.getInputStream(), fos);
-                return ResponseEntity.ok().build();
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+    @PostMapping(value = "/importRecipe", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
+    public ResponseEntity<String> uploadDataFileRecipe(@RequestParam MultipartFile file) {
+        filesRecipeService.cleanDataFile();
+        File dataFile = filesRecipeService.getDataFile();
+        try (FileOutputStream fos = new FileOutputStream(dataFile)) {
+            IOUtils.copy(file.getInputStream(), fos);
+            return ResponseEntity.ok().build();
+        } catch (IOException e) {
+            e.printStackTrace();
         }
-
-        @PostMapping(value = "/importIngredient", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-        public ResponseEntity<String> uploadDataFileIngredient (@RequestParam MultipartFile file){
-            filesIngredientService.cleanDataFile();
-            File dataFile = filesIngredientService.getDataFile();
-            try (FileOutputStream fos = new FileOutputStream(dataFile)) {
-                IOUtils.copy(file.getInputStream(), fos);
-                return ResponseEntity.ok().build();
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
-        }
+        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
     }
+
+    @PostMapping(value = "/importIngredient", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
+    public ResponseEntity<String> uploadDataFileIngredient(@RequestParam MultipartFile file) {
+        filesIngredientService.cleanDataFile();
+        File dataFile = filesIngredientService.getDataFile();
+        try (FileOutputStream fos = new FileOutputStream(dataFile)) {
+            IOUtils.copy(file.getInputStream(), fos);
+            return ResponseEntity.ok().build();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+    }
+}
