+++ src/main/java/me/matvienkoeg/recipeapp/services/impl/FilesRecipeServiceImpl.java
@@ -1,7 +1,5 @@
 package me.matvienkoeg.recipeapp.services.impl;
 
-import me.matvienkoeg.recipeapp.model.Ingredient;
-import me.matvienkoeg.recipeapp.model.Recipe;
 import me.matvienkoeg.recipeapp.services.FilesRecipeService;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
@@ -10,8 +8,6 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.Map;
-import java.util.TreeMap;
 
 @Service
 public class FilesRecipeServiceImpl implements FilesRecipeService {
@@ -22,7 +18,20 @@ public class FilesRecipeServiceImpl implements FilesRecipeService {
     @Value("${name.of.recipe.txt.file}")
     private String recipesTxtFileName;
 
-    private Map<Long, Recipe> recipes = new TreeMap<>();
+    @Override
+    public String getDataFilePath() {
+        return dataFilePath;
+    }
+
+    @Override
+    public String getDataFileName() {
+        return dataFileName;
+    }
+
+    @Override
+    public String getRecipesTxtFileName() {
+        return recipesTxtFileName;
+    }
 
     @Override
     public boolean saveToFile(String json) {
@@ -56,10 +65,6 @@ public class FilesRecipeServiceImpl implements FilesRecipeService {
         return Files.writeString(path, content);
     }
 
-    @Override
-    public File prepareRecipesTxt() throws IOException {
-        return saveToTxt (recipesToString(), Path.of(dataFilePath, recipesTxtFileName)).toFile();
-    }
 
     @Override
     public boolean cleanDataFile() {
@@ -72,23 +77,6 @@ public class FilesRecipeServiceImpl implements FilesRecipeService {
             e.printStackTrace();
             return false;
         }
-
-    }
-    public  String  recipesToString() {
-        StringBuilder sb = new StringBuilder();
-        String listEl = " * ";
-        for (Recipe recipe : recipes.values()) {
-            sb.append("\n").append(recipe.toString()).append("\n");
-            sb.append("\nИнгредиенты:\n");
-            for (Ingredient ingredient : recipe.getIngredients()) {
-                sb.append(listEl).append(ingredient.toString()).append("\n");
-            }
-            sb.append("\nИнструкция приготовления:\n");
-            for (String сookingInstruction : recipe.getCookingInstructions()) {
-                sb.append(listEl).append(сookingInstruction).append("\n");
-            }
-        }
-        return sb.append("\n").toString();
 
     }
