+++ src/main/resources/recipes.txt
@@ -1,1 +1,43 @@
 
+Сырники из Tворога
+ cookingTime: 30
+ cookingTimeUnitMeasurement: мин.
+ numberServings: 2
+ numberServingsUnitMeasurement: шт.
+
+Ингредиенты:
+ * Творог - 350 гр.
+ * Куриное яйцо - 2 шт.
+ * Пшеничная мука - 6 ст. л.
+ * Сахар - 2 ст. л.
+
+Инструкция приготовления:
+ * Смешайте весь творог с яйцами, сахаром и тщательно всё перемешайте.
+ * Всыпьте в творог муку и тщательно перемешайте.
+ * Поставьте сковороду на средний огонь и налейте в нее подсолнечное масло.
+ * Слепите несколько небольших шариков из получившейся творожной массы и положите их на тарелку. Затем по очереди обкатывайте творожные шарики в муке и выкладывайте на сковороду.
+ * Обжаривайте сырники 1–2 минуты до появления золотистой корочки. Затем переверните их на другую сторону и также обжарьте до золотистого состояния.
+ * Повторяйте, пока творог не закончится.
+
+Салат с креветками и кунжутом
+ cookingTime: 20
+ cookingTimeUnitMeasurement: мин.
+ numberServings: 2
+ numberServingsUnitMeasurement: шт.
+
+Ингредиенты:
+ * Помидоры  - 3 шт.
+ * Королевские креветки - 300 гр.
+ * Зеленый салат - 1 пучок
+ * Огурцы  - 2 шт.
+ * Специи - 0 по вкусу
+ * Майонез - 0 по вкусу
+ * Кунжутные семечки - 3 ст. л.
+
+Инструкция приготовления:
+ * Порезать овощи крупно, салат порвать на куски.
+ * Довести до кипения сильно соленую воду и добавить креветки. После кипения варить 5 минут.
+ * Очистить креветки от панциря.
+ * Смешать овощи и креветки.
+ * Посыпать кунжутными семечками и заправить майонезом.
+
